Skip builds if arch isn't x86_64 (where unsupported)

+++ scripts/phpspy_build.sh
@@ -5,6 +5,15 @@
 #
 set -euo pipefail
 
+# TODO support aarch64
+if [ $(uname -m) != "x86_64" ]; then
+    mkdir -p /phpspy
+    touch /phpspy/phpspy
+    mkdir -p /binutils/binutils-2.25/bin/bin/
+    touch /binutils/binutils-2.25/bin/bin/objdump
+    exit 0
+fi
+
 # First build phpspy
 pushd /
 git clone --depth=1 -b v0.6.0-g4 --recursive https://github.com/Granulate/phpspy.git && git -C phpspy reset --hard af1d74ebd4f9c27486e82397eafde1c450f06510
